+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Test for `--collisions`, `--suffix`, and `--parase` used at the same time.
+- Added `--dryrun`.
+- Added `.gif' suffix.
+- List of file and directory names to ignore.
+- Added suffix `.mov`.
+
+### Changed
+- `sort_images` first checks if the source image is a directory and skips it if true.
+- Suffix checked before the filename is parsed for a datetime stamp.
+
+### Fixed
+- Collisions not being populated.
+- Tests.
 
 ## [0.9.0] - 2020-02-13
 ### Added
@@ -65,7 +77,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - If a destination file exists and the hash of the source and destination files differ then skip it. If the
   hash is the same, then proceed (currently copy is called, but later it will be move so this will eliminate
-  the source file). 
+  the source file).
 
 ## [0.2.0] - 2020-02-09
 ### Added
@@ -76,7 +88,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] 2020-02-07
 ### Added
 - `search_directory` returns a list of the contents in a directory.
-- `is_file` to return True or False if something is a file. 
+- `is_file` to return True or False if something is a file.
 - `get_date_from_filename` for getting a `datetime.datetime` object from a filename.
 - Initial test images.
 - Script `resize.py` to help with prepping test images.